code_assistant/requirements.txt
openai>=1.3.0
python-dotenv>=0.19.0
pytest>=7.0.0
